Test dilation moving point filter
Detail:
If we want to use moving point filter in [1],
the error of first ICP result at perturbation must be small.
Compare 2 method:
1. grid, update (add-only) --> small error at perturbation
2. grid, no-update --> big error at perturbation
So I decide to use add-only update to compute the first ICP result.

Dilate wallcloud with 5*5 structure element.

When compare the dilation wallcloud and current frame,
there is a problem of "floating point numbers can not be
represented exactly in binary form" when apply "setdiff" or
"ismember". See:
http://matlab.wikia.com/wiki/FAQ#Why_is_0.3_-_0.2_-_0.1_.28or_similar.29_not_equal_to_zero.3F
Solve it by using "ismembertol" instead of the 2 previous functions.

Now I can pick up moving point with the first ICP result.

Data:
0801
	101017

Todo:
See if this method can be applied to other data well.